requirements.txt
tensorflow
pandas
pyyaml
dvc[ssh]
